Working version
- macos: make -j20 && make start
- FreeBSD:  make -j20 && make start

Check faxserver/system.conf & faxserver/faxout/*.conf for changes you want

Currently the faxserver/bin directory is not really clean

There are some strange glitches on building that might be because of my
homebrew or whatever. (it eventually tries to link against an unbuild x86-64 lib...  see Makefile for "hack"

URF to fax converter added.
In normal "mac-printing" cases, only urf files should be generated.
( urf ->  pgm -> sff -> fax )

If that is correct all the pdf conversion tools are not needed and only a small subset of netpbm binaries is used
(No imagemagic/poppler needs to be available..)