Target "make runserver" is all you need...
Currently the Printer shows as FB6591..  (but It most likly works for other FB too)

It can convert Universal raster files or PDF's (needs some tools installed for PDF) and send.

Verified working on a FreeBSD 13.1 installation and MacOS against a a FB6591 & macos 12 IPP printer driver
-- NOT working with ipados/ios -- At least the fax-part does not show, so no number entry ....

MacOS patches
see make_all_macos.sh for details

There are some strange glitches on building that might be because of my
homebrew or whatever.
On an M1 arm64 mac this should work